Add launcher with Screen for solving astor problem
Add launcher of DS using screen, which solves the problem
of that we have, trying to run the DS from Astor.

For an unknown reason we cannot start DS usint 'python ...'
from a DS launcher, and we cannot neither start the DS with
an entry point. Use screen as alternative while the server
is running on Suse, till we have the servers running on Debian.