Fix problem in DS finishing in ON state
The end command puts the DS in Standby, but the run_command working
in parallel, was putting afterwards the DS in ON state.
This was making the DS finishing in ON state, and, cause of that,
the txt file could not be set again.

Remove the set_state in run_command method; and adapt also the
code of magnetism end command, in order to be compatible with
this change.